refactor(kv-router): give CKF pools stable identities
