+++ lib/kv-router/src/lib.rs
@@ -10,6 +10,7 @@ mod active_set;
 pub(crate) mod cleanup;
 mod lookup_update;
 
+pub mod identity;
 pub mod indexer;
 pub mod protocols;
 pub mod recovery;
@@ -47,6 +48,7 @@ pub use config::{
     KvRouterConfig, RouterConfigOverride, RouterPrefillLoadModel, RouterQueuePolicy,
     SharedCacheType,
 };
+pub use identity::DcId;
 #[allow(deprecated)]
 pub use indexer::{
     AnchorAwareBranchShardedIndexer, AnchorRef, AnchorTask, BranchShardedIndexer,
