+++ lib/kv-router/src/services/indexer/mod.rs
@@ -9,6 +9,11 @@
 //! a ZMQ listener that ingests its KV events into a per-(model, routing group)
 //! [`backend::Indexer`].
 //!
+//! `IndexerKey { model_name, routing_group }` remains this service's sole registry authority.
+//! Resolving it to `IndexerDomainId` is intentionally deferred until registration can carry
+//! authoritative explicit identity material; hashing local defaults here would create a second
+//! authority without enabling safe cross-service identity.
+//!
 //! ## Multi-tier responses
 //!
 //! `/query` and `/query_by_hash` return both:
