+++ lib/llm/src/model_card.rs
@@ -23,6 +23,7 @@ use crate::model_type::{ModelInput, ModelType};
 use crate::protocols::tensor::TensorModelConfig;
 use anyhow::{Context, Result};
 use derive_builder::Builder;
+use dynamo_kv_router::identity::{ExplicitIdentityMap, IndexerIdentitySpec};
 use dynamo_runtime::{slug::Slug, storage::kv};
 use serde::{Deserialize, Serialize};
 use tokenizers::Tokenizer as HfTokenizer;
@@ -31,6 +32,31 @@ use crate::preprocessor::media::{MediaDecoder, MediaFetcher};
 use crate::protocols::TokenIdType;
 
 const DEFAULT_TOKENIZER_CACHE_BYTES: usize = 64 * 1024 * 1024;
+
+fn append_indexer_identity_checksum(bytes: &mut Vec<u8>, spec: &IndexerIdentitySpec) {
+    bytes.extend_from_slice(b"dynamo/model-card/indexer-identity/v1");
+    append_identity_dimension(bytes, spec.semantics());
+    append_identity_dimension(bytes, spec.routing_scope());
+}
+
+fn append_identity_dimension(bytes: &mut Vec<u8>, dimension: Option<&ExplicitIdentityMap>) {
+    let Some(dimension) = dimension else {
+        bytes.push(0);
+        return;
+    };
+    bytes.push(1);
+    bytes.extend_from_slice(&(dimension.entries().len() as u32).to_le_bytes());
+    for (key, value) in dimension.entries() {
+        append_framed_identity_value(bytes, key.as_bytes());
+        append_framed_identity_value(bytes, value.as_bytes());
+    }
+}
+
+fn append_framed_identity_value(bytes: &mut Vec<u8>, value: &[u8]) {
+    let len = u32::try_from(value.len()).expect("validated identity values fit u32");
+    bytes.extend_from_slice(&len.to_le_bytes());
+    bytes.extend_from_slice(value);
+}
 
 fn tokenizer_cache_enabled(value: Option<&str>) -> bool {
     !matches!(value, Some("0"))
@@ -905,6 +931,14 @@ pub struct ModelDeploymentCard {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub router_config: Option<RouterConfig>,
 
+    /// Optional authoritative KV-indexer compatibility and isolation material.
+    ///
+    /// A present dimension replaces its component-derived default. Entry labels are deliberately
+    /// uninterpreted so engine-specific compatibility facts do not expand this schema.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    #[builder(default)]
+    pub indexer_identity: Option<IndexerIdentitySpec>,
+
     /// Sibling files (e.g. `preprocessor_config.json`) the worker
     /// advertises alongside the typed slots.
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
@@ -1095,6 +1129,10 @@ impl ModelDeploymentCard {
                     // along with this effort, we should reorganize where RouterConfig
                     // should be defined.
                     bytes_to_hash.extend(blake3::hash(&bytes).as_bytes());
+                }
+
+                if let Some(identity) = self.indexer_identity.as_ref() {
+                    append_indexer_identity_checksum(&mut bytes_to_hash, identity);
                 }
 
                 // Aliases participate in the checksum. Every worker in a
@@ -1663,6 +1701,7 @@ impl ModelDeploymentCard {
             media_decoder: None,
             media_fetcher: None,
             router_config: None,
+            indexer_identity: None,
             extra_files: Vec::new(),
             checksum: OnceLock::new(),
         })
@@ -2942,6 +2981,7 @@ mod worker_type_tests {
 
     use super::*;
     use crate::worker_type::WorkerType;
+    use std::collections::BTreeMap;
 
     #[test]
     fn default_card_has_no_worker_type_and_no_needs() {
@@ -3081,5 +3121,37 @@ mod worker_type_tests {
         let back: ModelDeploymentCard = serde_json::from_str(&stripped).unwrap();
         assert_eq!(back.worker_type, None);
         assert!(back.needs.is_empty());
+    }
+
+    #[test]
+    fn indexer_identity_changes_mdcsum_and_missing_field_is_backward_compatible() {
+        fn spec(value: &str) -> IndexerIdentitySpec {
+            IndexerIdentitySpec::new(
+                Some(
+                    ExplicitIdentityMap::new(BTreeMap::from([(
+                        "weights".to_string(),
+                        value.to_string(),
+                    )]))
+                    .unwrap(),
+                ),
+                None,
+            )
+        }
+
+        let baseline = ModelDeploymentCard::with_name_only("model");
+        let mut explicit = ModelDeploymentCard::with_name_only("model");
+        explicit.indexer_identity = Some(spec("revision-a"));
+        assert_ne!(baseline.mdcsum(), explicit.mdcsum());
+
+        let mut serialized = serde_json::to_value(&explicit).unwrap();
+        assert!(
+            serialized
+                .as_object_mut()
+                .unwrap()
+                .remove("indexer_identity")
+                .is_some()
+        );
+        let restored: ModelDeploymentCard = serde_json::from_value(serialized).unwrap();
+        assert!(restored.indexer_identity.is_none());
     }
 }
